Support key rotation
To allow key rotation over time, introduce a new config key
`::private-keys`, a collection of keys to try building signatures
with.

The first key in the collection is always selected for signing,
but verifying tries all keys, this allows adding keys over time.

`sign` and `verify` now allow additional arities for the payload
and timestamp (or max-age).